+++ libs/pizza-parser/utils/config.ts
@@ -1,0 +1,1 @@
+export const HTTP_PROXY_URL = 'http://localhost:5566';