+++ libs/pizza-parser/utils/crypto.ts
@@ -1,0 +1,3 @@
+import crypto from 'crypto';
+
+export const toSha256 = (str: string) => crypto.createHash('sha256').update(str).digest('hex');